resources/requirements.txt
bibtexparser
python-slugify
python-frontmatter
pyyaml
pylev
requests
beautifulsoup4